pyproject.toml
[project]
name = "game-master-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.6.7",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
